refactor: improved types and component scopes
