+++ packages/web-app-notes/src/composables/stores/useTocStore.ts
@@ -10,13 +10,15 @@ export const sortTocNodes = (a: TocNode, b: TocNode) => {
 }
 
 export const useTocStore = defineStore('toc', () => {
-  const tocNodes = ref<TocNode[]>(null)
+  const tocNodes = ref<TocNode[] | null>(null)
 
   const setTocNodes = (nodes: TocNode[]) => {
     tocNodes.value = nodes
   }
   const addTocNode = (node: TocNode, parentNode?: TocNode) => {
     if (parentNode) {
+      // FIXME: we should find the parentNode in the tocNodes ref instead of mutating the parentNode
+      // (as it might be a copy of the original node)
       parentNode.children.push(node)
       parentNode.children.sort(sortTocNodes)
     } else {
@@ -29,7 +31,7 @@ export const useTocStore = defineStore('toc', () => {
     return unref(tocNodes) !== null
   })
 
-  const draggedNode = ref<TocNode>(null)
+  const draggedNode = ref<TocNode | null>(null)
   const setDraggedNode = (node: TocNode | null) => {
     draggedNode.value = node
     if (node === null) {
@@ -54,10 +56,17 @@ export const useTocStore = defineStore('toc', () => {
     dragOverRoot.value = value
   }
 
+  const toggleNodeCollapse = (node: TocNode) => {
+    // FIXME: we should find the node in the tocNodes ref instead of mutating the node
+    // (as it might be a copy of the original node)
+    node.collapsed = !node.collapsed
+  }
+
   const clearTocNodes = () => {
     tocNodes.value = null
     draggedNode.value = null
     dragOverNodeId.value = null
+    dragOverRoot.value = false
   }
 
   return {
@@ -74,6 +83,7 @@ export const useTocStore = defineStore('toc', () => {
     setDragOverNodeId,
     isDragOverNode,
     dragOverRoot,
-    setDragOverRoot
+    setDragOverRoot,
+    toggleNodeCollapse
   }
 })
